requirements.txt
future==0.16.0
pdfminer==20140328
pefile==2016.3.28
Pillow==3.4.2
PyInstaller==3.2
pyocr==0.4.2
pypiwin32==219
six==1.10.0
